refactor(mcp): extract ToolCall to separate model file #371
Move ToolCall data class and parsing logic from McpChatResultPanel to a dedicated model file.
This improves code organization and separation of concerns.
